Make button and form field to start, finish and submit trip (#6)
**Feature changes:**
- Allows to "start a trip" via the click of the button. At the moment
that does nothing but change the color of the button and show a
different button text. In the future this is where we would start the
GPS tracking from.
- Allows to stop the trip. 
- After stopping the trip, shows a text field for entering some trip
notes. These can then be "submitted" by the click of a button. Nothing
happens here yet, but it is the base for building the trip object with
the gps and notes data, and sending it via POST call to the .NET
backend.
- The view then changes back to the start view. 

- Adds a test: I am aware this uses a deprecated method, but i wasn't
able to make this work otherwies. I added a refactoring ticket to the
backlog for android Trello bug card
https://trello.com/c/XuNCEJFS/57-android-bug-backlog

**DevOps changes:**
Adds a test runner for Github. It will now run all tests when a new PR
is made, to any branch.