Track live GPS (#13)
This PR introduces the actual GPS Tracking.

1) It creates a TrackingService. Its role is to track and broadcast
location updates. It creates a foreground service, and attaches the GPS
tracking to it. We could also run this without the foregound service,
but then the tracking would not work when the phone sleeps. With this
implementation we can track the GPS also when the phone sleeps, which is
100% essential.

2) This PR generally restructures the folder structure. This was needed
as we had too many files in random places.

A follow up ticket that I want to explore in the future is here:

https://trello.com/c/9IqQd1jK/138-explore-if-it-is-possible-to-give-the-user-a-visual-clue-the-android-app-is-runnin-in-the-top-app-bar-show-foregroundservice