Make code more modular (#22)
This PR seeks to make the code more modular and testable. I'm reducin
the TrackingFragment class from 300 to around 220 lines of code, and I
am adding more tests.
There are some classes that I tried to test, but I just don't have
enough testing knowledge for the tests that interact closely with the
Android system.

Note: some of these tests will only work, if you as a tester actively
interact with the mock device on your android screen (i.e. giving
permissions when asked on screen).